third_party/blink/web_tests/shapedetection/detector-same-object.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/big_buffer.mojom.js"></script>
<script src="file:///gen/services/shape_detection/public/mojom/textdetection.mojom.js"></script>
<script src="resources/big-buffer-helpers.js"></script>
<script src="resources/mock-textdetection.js"></script>
<script>

var createTestForImageData = function(createDetector, mock,
                                      detectionResultTest) {
  async_test(function(t) {
    var img = new Image();
    img.onload = async function() {

      var canvas = document.createElement("canvas");;
      canvas.getContext("2d").drawImage(img, 0, 0);

      var detector = createDetector();
      try {
        var detectionResult = await detector.detect(canvas.getContext("2d")
            .getImageData(0, 0, canvas.width, canvas.height));
        detectionResultTest(detectionResult);
        t.done();
      } catch(error) {
        assert_unreached("Error during detect(canvas): " + error);
      }
    }

    img.src = "../media/content/greenbox.png";
  });
};

function CheckDetectedTextBlocksSameObjects(detectedTextBlocks) {
  assert_greater_than(detectedTextBlocks.length, 0, "Number of textBlocks");
  assert_equals(detectedTextBlocks[0].rawValue, detectedTextBlocks[0].rawValue);
  assert_equals(detectedTextBlocks[0].boundingBox, detectedTextBlocks[0].boundingBox);
}

// These tests verify that detect()ed Detected Text's individual
// fields are [SameObject].
generate_tests(createTestForImageData, [
  [
    "Text - detect(ImageData), [SameObject]",
    () => { return new TextDetector(); },
    mockTextDetection,
    CheckDetectedTextBlocksSameObjects
  ]
]);

</script>
